Announce one meetup at a time
Split the groups up to make one API request per group. Make a limit of 1 future event. If the response comes back as an empty array, store a '-' in robot.brain for that particular meetup. Otherwise, get the meetup's ID, check it against robot.brain and if it's different, announce it.